Revert "Adds UI actions for all the supported ISM actions (#49)"
This reverts commit 0d6b814a14ec8314325050c81b9d87946656f098.
