--- public/pages/VisualCreatePolicy/components/UIActions/index.ts
@@ -1,40 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- *
- * Modifications Copyright OpenSearch Contributors. See
- * GitHub history for details.
- */
-
-import AllocationUIAction from "./AllocationUIAction";
-import CloseUIAction from "./CloseUIAction";
-import DeleteUIAction from "./DeleteUIAction";
-import ForceMergeUIAction from "./ForceMergeUIAction";
-import IndexPriorityUIAction from "./IndexPriorityUIAction";
-import NotificationUIAction from "./NotificationUIAction";
-import OpenUIAction from "./OpenUIAction";
-import ReadOnlyUIAction from "./ReadOnlyUIAction";
-import ReadWriteUIAction from "./ReadWriteUIAction";
-import ReplicaCountUIAction from "./ReplicaCountUIAction";
-import RolloverUIAction from "./RolloverUIAction";
-import RollupUIAction from "./RollupUIAction";
-import SnapshotUIAction from "./SnapshotUIAction";
-
-export {
-  AllocationUIAction,
-  CloseUIAction,
-  DeleteUIAction,
-  ForceMergeUIAction,
-  IndexPriorityUIAction,
-  NotificationUIAction,
-  OpenUIAction,
-  ReadOnlyUIAction,
-  ReadWriteUIAction,
-  ReplicaCountUIAction,
-  RolloverUIAction,
-  RollupUIAction,
-  SnapshotUIAction,
-};
