+++ public/pages/VisualCreatePolicy/utils/helpers.ts
@@ -9,38 +9,7 @@
  * GitHub history for details.
  */
 
-import { UIAction, Action, Transition } from "../../../../models/interfaces";
-import {
-  ActionType,
-  DEFAULT_ALLOCATION,
-  DEFAULT_CLOSE,
-  DEFAULT_DELETE,
-  DEFAULT_FORCE_MERGE,
-  DEFAULT_INDEX_PRIORITY,
-  DEFAULT_NOTIFICATION,
-  DEFAULT_OPEN,
-  DEFAULT_READ_ONLY,
-  DEFAULT_READ_WRITE,
-  DEFAULT_REPLICA_COUNT,
-  DEFAULT_ROLLOVER,
-  DEFAULT_ROLLUP,
-  DEFAULT_SNAPSHOT,
-} from "./constants";
-import {
-  AllocationUIAction,
-  CloseUIAction,
-  DeleteUIAction,
-  ForceMergeUIAction,
-  IndexPriorityUIAction,
-  NotificationUIAction,
-  OpenUIAction,
-  ReadOnlyUIAction,
-  ReadWriteUIAction,
-  ReplicaCountUIAction,
-  RolloverUIAction,
-  RollupUIAction,
-  SnapshotUIAction,
-} from "../components/UIActions";
+import { Transition } from "../../../../models/interfaces";
 
 export const getConditionContent = (transition: Transition): string => {
   const {
@@ -56,44 +25,4 @@ export const getConditionContent = (transition: Transition): string => {
   if (minIndexAge != undefined) return `Minimum index age is ${minIndexAge}`;
   if (cron != undefined) return `After cron expression "${cron.cron.expression}" in ${cron.cron.timezone}`;
   return "No condition";
-};
-
-export const getUIActionFromData = (action: Action): UIAction<any> => {
-  const actionType = Object.keys(action).pop();
-  if (!actionType) throw new Error(`Failed to get action using type [${actionType}]`);
-  const uiAction = getUIAction(actionType);
-  return uiAction.clone(action);
-};
-
-export const getUIAction = (actionType: string): UIAction<any> => {
-  switch (actionType) {
-    case ActionType.Allocation:
-      return new AllocationUIAction(DEFAULT_ALLOCATION);
-    case ActionType.Close:
-      return new CloseUIAction(DEFAULT_CLOSE);
-    case ActionType.Delete:
-      return new DeleteUIAction(DEFAULT_DELETE);
-    case ActionType.ForceMerge:
-      return new ForceMergeUIAction(DEFAULT_FORCE_MERGE);
-    case ActionType.IndexPriority:
-      return new IndexPriorityUIAction(DEFAULT_INDEX_PRIORITY);
-    case ActionType.Notification:
-      return new NotificationUIAction(DEFAULT_NOTIFICATION);
-    case ActionType.Open:
-      return new OpenUIAction(DEFAULT_OPEN);
-    case ActionType.ReadOnly:
-      return new ReadOnlyUIAction(DEFAULT_READ_ONLY);
-    case ActionType.ReadWrite:
-      return new ReadWriteUIAction(DEFAULT_READ_WRITE);
-    case ActionType.ReplicaCount:
-      return new ReplicaCountUIAction(DEFAULT_REPLICA_COUNT);
-    case ActionType.Rollover:
-      return new RolloverUIAction(DEFAULT_ROLLOVER);
-    case ActionType.Rollup:
-      return new RollupUIAction(DEFAULT_ROLLUP);
-    case ActionType.Snapshot:
-      return new SnapshotUIAction(DEFAULT_SNAPSHOT);
-    default:
-      throw new Error(`Action type [${actionType}] not supported`);
-  }
 };
